Complete up to policy selection
A lot of refactoring was done in this update. Bot action now works properly and is its own file, representing the automated gameplay for bots. Verification for commands is in a file  which makes it easier to see which commands can run at what time. General gameplay is now functioning as intended up to the policy selection - the next phase is enacting/vetoing a policy as a Chancellor. Additionally, the server info has been added to DMs.