Added image support
After finding the assets on secrethitler.io, I used Pillow to generate images that are used in embeds. Now, the embed constructing method send_message() takes a parameter image which is the file name for the image (searches asset folder for game). Images are added for role assignment and policy selection, as well as a more informative popup when using $load.