Game functions complete, emoji bank for GUI
The game (untested) has all of the core features implemented. Policies can be chosen using reacts and there are special reactions for veto. An emoji vault server was created to house some of the icons that will be used for a player GUI system, allowing players to join and leave by clicking the reactions.