Create board image, add player join using reactions
All instances of using PIL Image instances were deferred to bytearrays (io.BytesIO) as it allows for uploading to Discord without any saving of images locally. This allowed creation of the board progress image, which merges the liberal and fascist boards. Support for election tracker will be added next. The GUI for player join features reactions to join, leave, add n bots, and start the game if conditions are met. The only things left untested are veto actions - full gameplay tests are the next step.